Well now you can call the Java code (See description for details)
Persistence package has been logged out of the compile. Check compilation to reenter. 
Node.Js calls to the Java code should now sort of work - still needs to be finished with implementation so that the remaining functions get called correctly, but if you call the /startProduction you can call on Daddy Svane <3