Added temporary work on Node API - Began the change to Spring Boot
The temporary work/attempts at solving the issues with the Node API have been added. However, due to issues with the Java-backend communicating back to the Node API, a transition to a Spring Boot API is being made. It does require a whole lot of restructuring the Domain, so it is nowhere near ready - we be working on it :)