Not Functional: Initial Springboot setup (See description)
Added a new directory in the Java directory called SpringbootApp, which contains all the changes to the domain to make it Springboot-compatible; it contains a copy of each class from the original Domain package, and the initial attempt at making it compatible, by making the classes @Entity and setting the Id. However, it is not finished, does not work.