Fixed a few datatypes in nodeOPCUA and added a connection from API
A few datatypes in the nodes that need to be written to the OPC UA server was in the wrong format. They are now changed to the right format. 

Added a connection from the API routes to the OPC UA Client. The API should now be able to call a function that will start a connection to the beer production machine and start a production(A dummy production so far, more will be implemented later)