Added more dynamic data to SimulationGraph
Pushed before adding jQuery
Implemented optimal function so it's drawn and added a errorMargin framework.
There is still alot of fine tuning needed and how the graph is shown must be agreed as it has a big influence on how the code is written.

Added a StartProduction javascript file which should handle HTTP POST request to the API. It's run from CreateProductionForm