Added function to get the production count to the machine API
Route, controller and machine function added, so ready to be used. 
Will return the results as a json object. Will only return usefull numbers if the machine has succesfully completed a production and is in state 17. 
If it isnt then it will return null as value. 
It returns the count for both defective and accectable products.