Optimization calculations set up. Needs testing
A total of 5 optimization calculation functions can now be called. These are: 
1. CalculateErrorSpeed, which takes in a batch size and an accepted error margin (in %) to produce a maximum speed that the user can run the machine at, without exceeding the acceptable error margin. 
2. CalculateErrorMargin, which takes in a speed to produce a minimum amount of errors that can be produced at that speed.
3. CalculateAmountOfErrors, which takes in a batch size and a margin, to calculate the amount of errors that the user produces for that batch, with the specified error margin
4. CalculateEstimatedProductionTime, which takes in a batch size and a speed, calculating the minimum amount of time that it will take to produce the batch
5. CalculateOptimalSpeed, which takes in a batch size and a time, and finds the minimum speed that the user will have to work with, to meet that deadline. 

The way that these informations are combined must make up the optimization estimations in the frontend. As such, it becomes possible to estimate several parameters, with the most important being:

The amount of errors for the optimal speed, which is found by:
Receiving the input of a batch size and the time allotted to the production and an acceptable error margin. These parameters can of course be found by making a combination of the previous formulas. 
IF THE USER allots a production time that provides a speed that produces a higher error margin than desired, the system should output an error (this must be implemented in the controller first however, see next commit) indicating whether or not this is a realistic goal.