Implemented check on whether or not the optimal speed is realistic
Now there is a check on the CalculateOptimalSpeed. Hence, the method now requires:
- batch size
- user-defined acceptable maximum error margin
- maximum time available
This calculates the speed that is needed to complete the batch in that allotted time, and uses this speed in the formula that was derived from the scatterplot. This will find the estimated errormargin that this speed will produce. 

The function then checks if the margin is bigger than the user-defined acceptable margin: if it is, then a 404 Bad Request message is returned, whereas if it is equal to or less than the specified margin, then the speed is returned.