Added finally and error messages to nodeOPCUA
Added the finally keyword in the try catch of every function that talks to the machine. This is to ensure that the connection(session) to the machine is closed down, as to not cause issued later on. 

Added more return functions so the functions now return something thats more relevant and will tell what happened.