Optimization for Errors implemented (See Description for details)
To access the optimization functions, the HTTP GET-request must be formatted: http://localhost:5000/Brewster/methodname
where "methodname" is one of the 4 routes that have been set up (calculateErrorSpeed, calculateErrorMargin, calculateEstimatedProductionTime, calculateOptimalSpeed) 

What's new: 

Comments have been added: it is now possible to explicitly see what a function expects to receive (See comment for @param req) and what the function will return (See comment for @param res) to the frontend

The calculateErrorSpeed function has been adjusted -  a parenthesis was put incorrectly, raising the wrong values to the power of 0.5

calculateAmountOfErrors has been commented out, since it is easier for the Minimum Viable Product to deal with the direct amount of acceptable errors than calculating percentages back and forth. 

The "calcMarg" variable has been renamed to "calcErrMarg"

Next up: An attempt at solving the calculation for returning correct beers to the user.