Updated Semantics + Calculate Valids now possible (See Description)
What's new:

Two new routes added and two routes were adjusted:
The routes /calculateAmountOfErrors and /calculatePercentageErrors have now been changed to be called /calculateAmountOfBeers and /calculatePercentageBeers. This is because these functions are universal for calculating both the amount of the batch size (that the percentage refers to) and the percentage (that the amount refers to) of the batch size is independent of whether or not the beers are valid or defective. 

The calculatePercentageErrors and calculateAmountOfErrors functions in Optimization controller had their names and comments updated.

The calculateValidMargin function was implemented. It takes a speed as input, and calculates how many valid beers that are produced at the given speed. It must therefore be passed a JSON object that determines the desired speed. 

The calculateValidSpeed function was implemented. It takes a margin as input, and calculates the speed (prior to x = 146.771) at which the specified amount of valid beers per minute are made.