Added new route + function (See description)
What's new:

Small adjustment to the brewster router which hadn't adjusted the function calls to the new names on calculateAmountOfBeers and calculatePercentageBeers.

Changed the name of the calculateOptimalSpeed route to calculateOptimalSpeedUsingErrors

Added new route to calculateOptimalSpeedUsingValids

Implemented new function in Optimization Controller that calculates whether or not the user's request for production speed is realistic; it takes the desired batch size (including errors) and the amount of beers of this batch size that has to be valid, as well as the amount of time that is as a maximum allotted to the production. From this, the function will calculate the minimum necessary speed, then it finds out how many valid beers per minute that the machine must make in order to meet the demand for valid beers (independent of time) and finally compares the amount of valid beers produced at the minimum necessary time to the specified desired amount of valid beers. 
If it is possible to produce exactly the desired amount of valid beers with the specified parameters, or more than the desired amount, within the given timeframe, then the function will return the speed that was calculated. 
If not, then an HTTP status code of 400 Bad Request will be sent to the frontend, as it is not possible to do.