Read the description. New try catch stuff added plus comments
What changes did I make? 
Moved some nodeID's variables from global to locally now that there only is 1 function that needed them. This was for change state, change command and currentstate. 
Added comments to all functions that describe what the function does. 
Added comments to most of the code to explain what the code does. 
Added a nested try catch statement in startsession
Added another thing to the json object that the exported function returns in case of an error is catched. 

Why did I make those changes? 
Comments are needed to document what the code does. Also so people have an easier time understanding the code. 

The nested try catch statement is added so the code will now try to connect to the simulation first, if it cant, an error should happen, that we then catch. We will then try to connect to the physical machine instead. 

The extra json object is just the error message that is given if it catch an error